+++ src/modules/datePipes.module.ts
@@ -1,0 +1,24 @@
+import {NgModule} from '@angular/core';
+
+import {DateConvertPipe} from '../pipes/dateConvert.pipe';
+import {DateFormatPipe} from '../pipes/dateFormat.pipe';
+
+/**
+ * Module used for exporting date pipes
+ */
+@NgModule(
+{
+    declarations:
+    [
+        DateFormatPipe,
+        DateConvertPipe
+    ],
+    exports:
+    [
+        DateFormatPipe,
+        DateConvertPipe
+    ]
+})
+export class DatePipesModule
+{
+}